Add basic (no pin) WPS function to Tweaks (#884)
* Add WPS function to Tweaks

## Adds:
- WPS connection ability through Tweaks -> Network

## Files added
- 3 icons (Need discussing)
- imgpop bin to show image on screen: https://github.com/XK9274/imgpop-miyoo
- SDL lib for imgpop

## Other changes
N/A

## Usage
- Press WPS button on router
- Press WPS connect button in tweaks
- Observe icon flashing red/orange indicating attempting to connect
- Observe icon flashing green indicating successfull connection
- Observe icon flashing red indicating failed connection

## Issues
- converting to draft to discuss icon/location/type of notifcation to user.

* 🎨 apply clang-format changes

* Bad merge, fix script.

* Add WPS to submenu

Added WPS into a submenu as I'll be adding WPS host into this same submenu. Fix some formatting.

* 🎨 apply clang-format changes

* Add a check to see if WIFI is enabled

Checks if the wlan0 interface exists/is enabled, if not runs AXP on and starts a supplicant/udhcpc process.

Shows new logos when querying wifi state

* Remove duped udhcpc kill

Remove a duped udhcpc kill that could cause blackscreen issues.

* Add sed -i to turn wifi on for MainUI

Tell MainUI that wifi needs to be kept up once started, if you dont do this it stays "off" in settings and the next time MainUI starts wifi is again turned off as the value is still 0.

* Fixed issue with wifi dropping after main relaunch

Issue caused by check_wifi in update_networking realising wifi set to 0 and killing wifi after using WPS.

Removed unnecessary loop

* Add if block to stop check_wifi killing wifi (see previous)

check_wifi realises Wifi was set to 0 in settings.json and kills wifi after relaunching MainUI (after a WPS connection)

* Add check for wpa_supp & udhcpc

Checks for wpa_supp and udhcpc before notifying that wifi is up

* rm not logical to have here

* Broken IF

Nothing to see here.....

* Sleep to show wifiup icon

---------

+++ static/build/.tmp_update/script/update_networking.sh
@@ -39,12 +39,15 @@ else
 			udhcpc -i wlan0 -s /etc/init.d/udhcpc.script &
 		fi
 	else
-		pkill -9 wpa_supplicant
-		pkill -9 udhcpc
-		/customer/app/axp_test wifioff
+		if [ ! -f "/tmp/dont_restart_wifi" ]; then
+			pkill -9 wpa_supplicant
+			pkill -9 udhcpc
+			/customer/app/axp_test wifioff
+		fi
 	fi
 fi
 }
+
 
 
 # Starts bftpd if the toggle is set to on
